Fixed #373 - Refactor MultiColSource.
