+++ django/db/models/sql/compiler.py
@@ -7,9 +7,8 @@ from itertools import chain
 from django.core.exceptions import EmptyResultSet, FieldError, FullResultSet
 from django.db import DatabaseError, NotSupportedError
 from django.db.models.constants import LOOKUP_SEP
-from django.db.models.expressions import F, OrderBy, RawSQL, Ref, Value
+from django.db.models.expressions import Cols, F, OrderBy, RawSQL, Ref, Value
 from django.db.models.fields.composite import (
-    Cols,
     CompositePrimaryKey,
     unnest_composite_fields,
 )
